v3: publish canonical briefs and Foundry contracts (1359 projects)

+++ v3/revert/src/AutoExit-2/lib/v3-core/contracts/interfaces/pool/IUniswapV3PoolState.sol
@@ -27,7 +27,7 @@ interface IUniswapV3PoolState {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint32 feeProtocol,
+            uint8 feeProtocol,
             bool unlocked
         );
